# Optimization batch 11: avoid repeatedly detecting same renames