… Documentation/technical/remembering-renames.txt
… t/t6429-merge-sequence-rename-caching.sh